Add a simple pyproject.toml
This allows automatic inclusion of build dependencies `setuptools` and
`Cython` when building pygame with `pip` or `build` frontends.

This pyproject.toml file is inefficient since `Cython` is only needed
for creating an sdist, but not for compiling it. The `pip` front end may
compile and install from an sdist on a user's computer if it cannot
download a suitable wheel; `Cython` will be downloaded but not used.