Unit tests work for subprocesses and test dir.
Make the unit test framework once again support running unit tests from
subprocesses, the *--usesubprocess* command line option, as well as
allow unit tests to be run from the test development directory. This is
accomplished by simplifying the imports required. The *test_utils*
submodule becomes a top level package. Many modules within the
*test_utils* directory also become top level modules.  In
*--usesubprocess* mode the test_runner.py script is now run directly as
a Python script rather than run as a submodule using the Python *-m*
command line option.

A problem is that the test_timer_common_reference test in *time_test.py*
fails when *time_test.py* is run in a subprocess, but not otherwise.
Until the reason is determined this patch will be put on hold. It may be
the *--usesubprocess* option is not repairable.

This is only a partial repair. Unit test modules *draw_test.py*,
*ftfont_test.py*, image__save_gl_*surface_test.py*, *rect_test.py*,
*surface_test.py* and *transform_test.py* still need their imports
updated to the new module path arrangement. These updates will be made
only if the *time_test.py* issue is resolved.